sports_booking/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
gunicorn
drf-yasg
